#pygrain Fix segfault when starting multiple mp_prefetches concurrently.
Parallel calls to `start_prefetch` result in concurrent calls to `SharedMemoryArray.enable_async_del`, which attempt to concurrently modify class-level state. Before the fix, the added unit test segfaults about 10% of the time, and the repro in experimental/ segfaults consistently. Using a higher number of iterators in the unit test results in forge OOMs. After the fix, the test passes with --runs_per_test=1000

PiperOrigin-RevId: 712578730