#grain Simplify state management in FirstFitPackingIterator.
The conditional in get_state() was making the class hard to reason about.

Now we have a single field for holding the checkpointed parent state. We advance it whenever a batch is fully emitted, since at that point the parent will no longer need to be rewound to before the batch.

PiperOrigin-RevId: 720967916